fixing memory leak in DragDropContext
